Refactoring of classes name
