+++ src/main/java/fr/uge/chargepointconfiguration/chargepoint/ocpp/OcppMessageParser.java
@@ -2,7 +2,7 @@ package fr.uge.chargepointconfiguration.chargepoint.ocpp;
 
 import fr.uge.chargepointconfiguration.chargepoint.WebSocketRequestMessage;
 import fr.uge.chargepointconfiguration.chargepoint.ocpp.ocpp16.OcppMessageParser16;
-import fr.uge.chargepointconfiguration.chargepoint.ocpp.ocpp2.OcppMessageParser2;
+import fr.uge.chargepointconfiguration.chargepoint.ocpp.ocpp2.OcppMessageParser20;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -26,7 +26,7 @@ public interface OcppMessageParser {
     Objects.requireNonNull(ocppVersion);
     return switch (ocppVersion) {
       case V1_6 -> new OcppMessageParser16();
-      case V2 -> new OcppMessageParser2();
+      case V2 -> new OcppMessageParser20();
     };
   }
 }
